chore: rename parse-driver folder to drivers and create new folder to MongoDB
